requirement.txt
flask==3.0.0
flask-restful==0.3.10
flask-jwt-extended==4.6.0
flask-cors==4.0.0
flask-bcrypt==1.0.1
pymongo==4.6.1
pika==1.3.2
boto3==1.34.0
python-dotenv==1.0.0
marshmallow==3.20.2
celery==5.3.4
redis==5.0.1